+++ tests/testthat/test_idf_object.R
@@ -288,3 +288,14 @@ test_that("S3 assigning works", {
     expect_silent(mat[["Roughness"]] <- "MediumSmooth")
     expect_equal(mat$Roughness, "MediumSmooth")
 })
+
+test_that("$possible_value() works", {
+    pos <- con$possible_value(c(3,1))
+    expect_is(pos, "IdfFieldPossible")
+    expect_equal(pos$field_index, c(3L, 1L))
+    expect_equal(pos$field_name, c("Layer 2", "Name"))
+    expect_equal(pos$auto, rep(NA_character_, 2L))
+    expect_equal(pos$default, rep(list(NA_character_), 2L))
+    expect_equal(pos$choice, rep(list(NA_character_), 2L))
+    expect_equal(pos$reference, list("NewMaterialName", NULL))
+})